v0.82.0: 新增factor_store_v5、tail_v5_pipeline模块，更新并行计算及流水线相关代码
新增:
- src/factor_store_v5.rs: 新版因子存储模块
- src/tail_v5_pipeline.rs: v5尾部流水线处理
- examples/: 示例代码

修改:
- Cargo.lock/Cargo.toml: 依赖更新
- src/lib.rs: 注册新模块
- src/backup_reader.rs: 备份读取器更新
- src/bin/worker_pipeline.rs: worker流水线更新
- src/factor_pipeline.rs: 因子流水线优化
- src/fast_csv_reader.rs: CSV读取器优化
- src/features.rs: 特征模块更新
- src/observable_order_metrics.rs: 订单指标更新
- src/order_pair_metrics_pipeline.rs: 订单对指标流水线
- src/parallel_computing.rs: 并行计算优化
- src/parallel_computing_date_only.rs: 日期并行计算
- *.pyi: 类型声明同步更新

+++ src/order_pair_metrics_pipeline.rs
@@ -21,10 +21,10 @@ fn array_to_fast_records(trades: ArrayView2<f64>) -> Vec<fast_csv_reader::TradeR
     let mut out = Vec::with_capacity(n);
     for i in 0..n {
         out.push(fast_csv_reader::TradeRecord {
-            time_sec: trades[[i, 0]],
-            price: trades[[i, 1]],
-            volume: trades[[i, 2]],
-            turnover: trades[[i, 3]],
+            time_sec: trades[[i, 0]] as f32,
+            price: trades[[i, 1]] as f32,
+            volume: trades[[i, 2]] as f32,
+            turnover: trades[[i, 3]] as f32,
             flag: trades[[i, 4]] as i32,
             bid_order: trades[[i, 5]] as i64,
             ask_order: trades[[i, 6]] as i64,
@@ -385,10 +385,10 @@ fn parse_trades_from_fast(records: &[fast_csv_reader::TradeRecord]) -> Vec<Trade
     records
         .iter()
         .map(|r| TradeRecord {
-            time: r.time_sec,
-            price: r.price,
-            volume: r.volume,
-            turnover: r.turnover,
+            time: r.time_sec as f64,
+            price: r.price as f64,
+            volume: r.volume as f64,
+            turnover: r.turnover as f64,
             flag: r.flag,
             bid_order: r.bid_order,
             ask_order: r.ask_order,
